Split off ossl signature abstrations in a module
