+++ ossl/src/lib.rs
@@ -21,6 +21,8 @@ use std::ffi::{c_char, c_int, c_long, c_uchar, c_uint, c_ulong, c_void, CStr};
 
 use crate::bindings::*;
 
+pub mod signature;
+
 #[cfg(feature = "fips")]
 pub mod fips;
 
@@ -142,11 +144,12 @@ macro_rules! trace_ossl {
                 file!(),
                 line!(),
                 $name,
-                ossl_err_stack()
+                crate::ossl_err_stack()
             );
         }
     };
 }
+pub(crate) use trace_ossl;
 
 /// A structure representing the main crypto library context
 pub struct OsslContext {
@@ -1316,1082 +1319,3 @@ impl Drop for EvpPkey {
 
 unsafe impl Send for EvpPkey {}
 unsafe impl Sync for EvpPkey {}
-
-/// Wrapper around OpenSSL's `EVP_SIGNATURE`, used for ML-DSA operations.
-pub struct EvpSignature {
-    ptr: *mut EVP_SIGNATURE,
-}
-
-impl EvpSignature {
-    /// Creates a new `EvpSignature` instance by fetching it by name.
-    pub fn new(ctx: &OsslContext, name: &CStr) -> Result<EvpSignature, Error> {
-        let ptr: *mut EVP_SIGNATURE = unsafe {
-            EVP_SIGNATURE_fetch(ctx.ptr(), name.as_ptr(), std::ptr::null_mut())
-        };
-        if ptr.is_null() {
-            trace_ossl!("EVP_SIGNATURE_fetch()");
-            return Err(Error::new(ErrorKind::NullPtr));
-        }
-        Ok(EvpSignature { ptr })
-    }
-
-    /// Returns a const pointer to the underlying `EVP_SIGNATURE`.
-    #[allow(dead_code)]
-    pub fn as_ptr(&self) -> *const EVP_SIGNATURE {
-        self.ptr
-    }
-
-    /// Returns a mutable pointer to the underlying `EVP_SIGNATURE`.
-    pub fn as_mut_ptr(&mut self) -> *mut EVP_SIGNATURE {
-        self.ptr
-    }
-}
-
-impl Drop for EvpSignature {
-    fn drop(&mut self) {
-        unsafe {
-            EVP_SIGNATURE_free(self.ptr);
-        }
-    }
-}
-
-macro_rules! cstr {
-    ($str:expr) => {
-        unsafe { CStr::from_ptr($str.as_ptr() as *const c_char) }
-    };
-}
-
-/// Known algorithms selectable for OsslSignature
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum SigAlg {
-    Ecdsa,
-    EcdsaSha1,
-    EcdsaSha2_224,
-    EcdsaSha2_256,
-    EcdsaSha2_384,
-    EcdsaSha2_512,
-    EcdsaSha3_224,
-    EcdsaSha3_256,
-    EcdsaSha3_384,
-    EcdsaSha3_512,
-    Ed25519,
-    Ed25519ctx,
-    Ed25519ph,
-    Ed448,
-    Ed448ph,
-    Mldsa44,
-    Mldsa65,
-    Mldsa87,
-    Rsa,
-    RsaNoPad,
-    RsaSha1,
-    RsaSha2_224,
-    RsaSha2_256,
-    RsaSha2_384,
-    RsaSha2_512,
-    RsaSha3_224,
-    RsaSha3_256,
-    RsaSha3_384,
-    RsaSha3_512,
-    RsaPss,
-    RsaPssSha1,
-    RsaPssSha2_224,
-    RsaPssSha2_256,
-    RsaPssSha2_384,
-    RsaPssSha2_512,
-    RsaPssSha3_224,
-    RsaPssSha3_256,
-    RsaPssSha3_384,
-    RsaPssSha3_512,
-}
-
-/// Helper that indicates if a signature algorithm should use oneshot apis
-fn sigalg_is_oneshot(alg: SigAlg) -> bool {
-    match alg {
-        SigAlg::Ecdsa | SigAlg::Rsa | SigAlg::RsaPss | SigAlg::RsaNoPad => true,
-        _ => false,
-    }
-}
-
-/// Helper that indicates if a signature algorithm should use legacy apis
-fn sigalg_uses_legacy_api(alg: SigAlg) -> bool {
-    #[cfg(not(ossl_v350))]
-    match alg {
-        SigAlg::Ecdsa | SigAlg::Rsa | SigAlg::RsaPss | SigAlg::RsaNoPad => {
-            false
-        }
-        SigAlg::Mldsa44 | SigAlg::Mldsa65 | SigAlg::Mldsa87 => false,
-        _ => true,
-    }
-    #[cfg(ossl_v350)]
-    match alg {
-        SigAlg::RsaPssSha1
-        | SigAlg::RsaPssSha2_224
-        | SigAlg::RsaPssSha2_256
-        | SigAlg::RsaPssSha2_384
-        | SigAlg::RsaPssSha3_224
-        | SigAlg::RsaPssSha3_256
-        | SigAlg::RsaPssSha3_384
-        | SigAlg::RsaPssSha3_512 => true,
-        _ => false,
-    }
-}
-
-/// Helper that indicates is a signature algorithm supports updates
-/// a value of None indicates autodetection should be performed
-fn sigalg_supports_updates(alg: SigAlg) -> Option<bool> {
-    match alg {
-        SigAlg::Ecdsa
-        | SigAlg::Ed25519
-        | SigAlg::Ed25519ctx
-        | SigAlg::Ed25519ph
-        | SigAlg::Ed448
-        | SigAlg::Ed448ph
-        | SigAlg::Rsa
-        | SigAlg::RsaPss
-        | SigAlg::RsaNoPad => Some(false),
-        SigAlg::Mldsa44 | SigAlg::Mldsa65 | SigAlg::Mldsa87 => None,
-        _ => Some(true),
-    }
-}
-
-static ECDSA_NAME: &CStr = c"ECDSA";
-static ECDSASHA1_NAME: &CStr = c"ECDSA-SHA1";
-static ECDSASHA2_224_NAME: &CStr = c"ECDSA-SHA2-224";
-static ECDSASHA2_256_NAME: &CStr = c"ECDSA-SHA2-256";
-static ECDSASHA2_384_NAME: &CStr = c"ECDSA-SHA2-384";
-static ECDSASHA2_512_NAME: &CStr = c"ECDSA-SHA2-512";
-static ECDSASHA3_224_NAME: &CStr = c"ECDSA-SHA3-224";
-static ECDSASHA3_256_NAME: &CStr = c"ECDSA-SHA3-256";
-static ECDSASHA3_384_NAME: &CStr = c"ECDSA-SHA3-384";
-static ECDSASHA3_512_NAME: &CStr = c"ECDSA-SHA3-512";
-static ED25519_NAME: &CStr = c"ED25519";
-static ED25519CTX_NAME: &CStr = c"ED25519CTX";
-static ED25519PH_NAME: &CStr = c"ED25519PH";
-static ED448_NAME: &CStr = c"ED448";
-static ED448PH_NAME: &CStr = c"ED448PH";
-static MLDSA44_NAME: &CStr = c"ML-DSA-44";
-static MLDSA65_NAME: &CStr = c"ML-DSA-65";
-static MLDSA87_NAME: &CStr = c"ML-DSA-87";
-static RSA_NAME: &CStr = c"RSA";
-static RSASHA1_NAME: &CStr = c"RSA-SHA1";
-static RSASHA2_224_NAME: &CStr = c"RSA-SHA2-224";
-static RSASHA2_256_NAME: &CStr = c"RSA-SHA2-256";
-static RSASHA2_384_NAME: &CStr = c"RSA-SHA2-384";
-static RSASHA2_512_NAME: &CStr = c"RSA-SHA2-512";
-static RSASHA3_224_NAME: &CStr = c"RSA-SHA3-224";
-static RSASHA3_256_NAME: &CStr = c"RSA-SHA3-256";
-static RSASHA3_384_NAME: &CStr = c"RSA-SHA3-384";
-static RSASHA3_512_NAME: &CStr = c"RSA-SHA3-512";
-/* The following names are not actually recognized by
- * OpenSSL and will cause a fetch error if used, they
- * have been made up for completeness, and debugging */
-static RSAPSS_NAME: &CStr = c"RSA-PSS";
-static RSAPSSSHA1_NAME: &CStr = c"RSA-PSS-SHA1";
-static RSAPSSSHA2_224_NAME: &CStr = c"RSA-PSS-SHA2-224";
-static RSAPSSSHA2_256_NAME: &CStr = c"RSA-PSS-SHA2-256";
-static RSAPSSSHA2_384_NAME: &CStr = c"RSA-PSS-SHA2-384";
-static RSAPSSSHA2_512_NAME: &CStr = c"RSA-PSS-SHA2-512";
-static RSAPSSSHA3_224_NAME: &CStr = c"RSA-PSS-SHA3-224";
-static RSAPSSSHA3_256_NAME: &CStr = c"RSA-PSS-SHA3-256";
-static RSAPSSSHA3_384_NAME: &CStr = c"RSA-PSS-SHA3-384";
-static RSAPSSSHA3_512_NAME: &CStr = c"RSA-PSS-SHA3-512";
-static RSANOPAD_NAME: &CStr = c"RSA-NO-PAD";
-
-/// Helper to return OpenSSL sigalg name
-fn sigalg_to_ossl_name(alg: SigAlg) -> &'static CStr {
-    match alg {
-        SigAlg::Ecdsa => ECDSA_NAME,
-        SigAlg::EcdsaSha1 => ECDSASHA1_NAME,
-        SigAlg::EcdsaSha2_224 => ECDSASHA2_224_NAME,
-        SigAlg::EcdsaSha2_256 => ECDSASHA2_256_NAME,
-        SigAlg::EcdsaSha2_384 => ECDSASHA2_384_NAME,
-        SigAlg::EcdsaSha2_512 => ECDSASHA2_512_NAME,
-        SigAlg::EcdsaSha3_224 => ECDSASHA3_224_NAME,
-        SigAlg::EcdsaSha3_256 => ECDSASHA3_256_NAME,
-        SigAlg::EcdsaSha3_384 => ECDSASHA3_384_NAME,
-        SigAlg::EcdsaSha3_512 => ECDSASHA3_512_NAME,
-        SigAlg::Ed25519 => ED25519_NAME,
-        SigAlg::Ed25519ctx => ED25519CTX_NAME,
-        SigAlg::Ed25519ph => ED25519PH_NAME,
-        SigAlg::Ed448 => ED448_NAME,
-        SigAlg::Ed448ph => ED448PH_NAME,
-        SigAlg::Mldsa44 => MLDSA44_NAME,
-        SigAlg::Mldsa65 => MLDSA65_NAME,
-        SigAlg::Mldsa87 => MLDSA87_NAME,
-        SigAlg::Rsa => RSA_NAME,
-        SigAlg::RsaSha1 => RSASHA1_NAME,
-        SigAlg::RsaSha2_224 => RSASHA2_224_NAME,
-        SigAlg::RsaSha2_256 => RSASHA2_256_NAME,
-        SigAlg::RsaSha2_384 => RSASHA2_384_NAME,
-        SigAlg::RsaSha2_512 => RSASHA2_512_NAME,
-        SigAlg::RsaSha3_224 => RSASHA3_224_NAME,
-        SigAlg::RsaSha3_256 => RSASHA3_256_NAME,
-        SigAlg::RsaSha3_384 => RSASHA3_384_NAME,
-        SigAlg::RsaSha3_512 => RSASHA3_512_NAME,
-        SigAlg::RsaPss => RSAPSS_NAME,
-        SigAlg::RsaPssSha1 => RSAPSSSHA1_NAME,
-        SigAlg::RsaPssSha2_224 => RSAPSSSHA2_224_NAME,
-        SigAlg::RsaPssSha2_256 => RSAPSSSHA2_256_NAME,
-        SigAlg::RsaPssSha2_384 => RSAPSSSHA2_384_NAME,
-        SigAlg::RsaPssSha2_512 => RSAPSSSHA2_512_NAME,
-        SigAlg::RsaPssSha3_224 => RSAPSSSHA3_224_NAME,
-        SigAlg::RsaPssSha3_256 => RSAPSSSHA3_256_NAME,
-        SigAlg::RsaPssSha3_384 => RSAPSSSHA3_384_NAME,
-        SigAlg::RsaPssSha3_512 => RSAPSSSHA3_512_NAME,
-        SigAlg::RsaNoPad => RSANOPAD_NAME,
-    }
-}
-
-/// Helper that returns the OpenSSL digest name associated to a sigalg
-fn sigalg_to_digest_ptr(alg: SigAlg) -> *const c_char {
-    match alg {
-        SigAlg::EcdsaSha1 | SigAlg::RsaSha1 | SigAlg::RsaPssSha1 => {
-            cstr!(OSSL_DIGEST_NAME_SHA1).as_ptr()
-        }
-        SigAlg::EcdsaSha2_224
-        | SigAlg::RsaSha2_224
-        | SigAlg::RsaPssSha2_224 => cstr!(OSSL_DIGEST_NAME_SHA2_224).as_ptr(),
-        SigAlg::EcdsaSha2_256
-        | SigAlg::RsaSha2_256
-        | SigAlg::RsaPssSha2_256 => cstr!(OSSL_DIGEST_NAME_SHA2_256).as_ptr(),
-        SigAlg::EcdsaSha2_384
-        | SigAlg::RsaSha2_384
-        | SigAlg::RsaPssSha2_384 => cstr!(OSSL_DIGEST_NAME_SHA2_384).as_ptr(),
-        SigAlg::EcdsaSha2_512
-        | SigAlg::RsaSha2_512
-        | SigAlg::RsaPssSha2_512 => cstr!(OSSL_DIGEST_NAME_SHA2_512).as_ptr(),
-        SigAlg::EcdsaSha3_224
-        | SigAlg::RsaSha3_224
-        | SigAlg::RsaPssSha3_224 => cstr!(OSSL_DIGEST_NAME_SHA3_224).as_ptr(),
-        SigAlg::EcdsaSha3_256
-        | SigAlg::RsaSha3_256
-        | SigAlg::RsaPssSha3_256 => cstr!(OSSL_DIGEST_NAME_SHA3_256).as_ptr(),
-        SigAlg::EcdsaSha3_384
-        | SigAlg::RsaSha3_384
-        | SigAlg::RsaPssSha3_384 => cstr!(OSSL_DIGEST_NAME_SHA3_384).as_ptr(),
-        SigAlg::EcdsaSha3_512
-        | SigAlg::RsaSha3_512
-        | SigAlg::RsaPssSha3_512 => cstr!(OSSL_DIGEST_NAME_SHA3_512).as_ptr(),
-        SigAlg::Mldsa44
-        | SigAlg::Mldsa65
-        | SigAlg::Mldsa87
-        | SigAlg::Ecdsa
-        | SigAlg::Ed25519
-        | SigAlg::Ed25519ctx
-        | SigAlg::Ed25519ph
-        | SigAlg::Ed448
-        | SigAlg::Ed448ph
-        | SigAlg::Rsa
-        | SigAlg::RsaPss
-        | SigAlg::RsaNoPad => std::ptr::null(),
-    }
-}
-
-pub struct RsaPssParams {
-    pub digest: &'static CStr,
-    pub mgf1: &'static CStr,
-    pub saltlen: c_int,
-}
-
-/// Helper to generate OsslParam arrays for initialization
-pub fn rsa_sig_params(
-    alg: SigAlg,
-    pss_params: &Option<RsaPssParams>,
-) -> Result<Option<OsslParam>, Error> {
-    match alg {
-        SigAlg::RsaNoPad => {
-            let mut params = OsslParam::new();
-            params.add_const_c_string(
-                cstr!(OSSL_SIGNATURE_PARAM_PAD_MODE),
-                cstr!(OSSL_PKEY_RSA_PAD_MODE_NONE),
-            )?;
-            params.finalize();
-            return Ok(Some(params));
-        }
-        SigAlg::Rsa
-        | SigAlg::RsaSha1
-        | SigAlg::RsaSha2_224
-        | SigAlg::RsaSha2_256
-        | SigAlg::RsaSha2_384
-        | SigAlg::RsaSha3_224
-        | SigAlg::RsaSha3_256
-        | SigAlg::RsaSha3_384
-        | SigAlg::RsaSha3_512 => {
-            /* In 3.5.0 there is direct sigalg support for
-             * PCKCS1 padding for digest algorithms so no
-             * paramters are needed or processed */
-            #[cfg(ossl_v350)]
-            if alg != SigAlg::Rsa {
-                return Ok(None);
-            }
-
-            let mut params = OsslParam::new();
-            params.add_const_c_string(
-                cstr!(OSSL_SIGNATURE_PARAM_PAD_MODE),
-                cstr!(OSSL_PKEY_RSA_PAD_MODE_PKCSV15),
-            )?;
-            params.finalize();
-            return Ok(Some(params));
-        }
-        SigAlg::RsaPss
-        | SigAlg::RsaPssSha1
-        | SigAlg::RsaPssSha2_224
-        | SigAlg::RsaPssSha2_256
-        | SigAlg::RsaPssSha2_384
-        | SigAlg::RsaPssSha3_224
-        | SigAlg::RsaPssSha3_256
-        | SigAlg::RsaPssSha3_384
-        | SigAlg::RsaPssSha3_512 => {
-            /* Pss always uses legacy interfaces, so we need
-             * all params set up */
-            if let Some(pss) = pss_params {
-                let mut params = OsslParam::new();
-
-                params.add_const_c_string(
-                    cstr!(OSSL_SIGNATURE_PARAM_PAD_MODE),
-                    cstr!(OSSL_PKEY_RSA_PAD_MODE_PSS),
-                )?;
-                params.add_const_c_string(
-                    cstr!(OSSL_SIGNATURE_PARAM_DIGEST),
-                    pss.digest,
-                )?;
-                params.add_const_c_string(
-                    cstr!(OSSL_SIGNATURE_PARAM_MGF1_DIGEST),
-                    pss.mgf1,
-                )?;
-                params.add_owned_int(
-                    cstr!(OSSL_SIGNATURE_PARAM_PSS_SALTLEN),
-                    pss.saltlen,
-                )?;
-                params.finalize();
-                return Ok(Some(params));
-            } else {
-                return Err(Error::new(ErrorKind::NullPtr));
-            }
-        }
-        _ => Err(Error::new(ErrorKind::WrapperError)),
-    }
-}
-
-/// Helper to generate OsslParam arrays for Eddsa initialization
-#[cfg(ossl_v320)]
-pub fn eddsa_params(
-    alg: SigAlg,
-    context: Option<Vec<u8>>,
-) -> Result<Option<OsslParam<'static>>, Error> {
-    match alg {
-        SigAlg::Ed25519
-        | SigAlg::Ed25519ph
-        | SigAlg::Ed448
-        | SigAlg::Ed448ph
-        | SigAlg::Ed25519ctx => (),
-        _ => return Err(Error::new(ErrorKind::WrapperError)),
-    }
-
-    /* With 3.5.0 we use the sigalg interface so no need
-     * for params */
-    #[cfg(ossl_v350)]
-    if context.is_none() {
-        return Ok(None);
-    }
-
-    let mut params = OsslParam::new();
-
-    #[cfg(not(ossl_v350))]
-    params.add_const_c_string(
-        cstr!(OSSL_SIGNATURE_PARAM_INSTANCE),
-        sigalg_to_ossl_name(alg),
-    )?;
-
-    if let Some(v) = context {
-        params.add_owned_octet_string(
-            cstr!(OSSL_SIGNATURE_PARAM_CONTEXT_STRING),
-            v,
-        )?;
-    }
-
-    params.finalize();
-    return Ok(Some(params));
-}
-
-/// Maximum buffer size for accumulating data when emulating multi-part
-/// operations for OpenSSL versions that only support one-shot operations.
-const MAX_BUFFER_LEN: usize = 1024 * 1024;
-
-/// Higher level wrapper for signature operations with OpenSSL
-#[derive(Debug)]
-pub struct OsslSignature {
-    /// The underlying OpenSSL EVP PKEY context.
-    pkey_ctx: EvpPkeyCtx,
-    /// The MD Ctx for cases when the old EVP_Digest interfaces need to be used
-    #[cfg(not(feature = "fips"))]
-    legacy_ctx: Option<EvpMdCtx>,
-    #[cfg(feature = "fips")]
-    legacy_ctx: Option<fips::ProviderSignatureCtx>,
-    /// Flag indicating if the current OpenSSL version supports multi-part
-    /// updates.
-    supports_updates: bool,
-    /// Buffer to accumulate data for multi-part emulation if needed.
-    data: Option<Vec<u8>>,
-    /// Stored signature for VerifySignature operations if updates aren't
-    /// supported by the OpenSSL provider.
-    signature: Option<Vec<u8>>,
-}
-
-impl OsslSignature {
-    /// Internal function to init the structure from a key
-    fn init(
-        ctx: &OsslContext,
-        key: &mut EvpPkey,
-    ) -> Result<OsslSignature, Error> {
-        Ok(OsslSignature {
-            pkey_ctx: key.new_ctx(ctx)?,
-            legacy_ctx: None,
-            supports_updates: false,
-            data: None,
-            signature: None,
-        })
-    }
-
-    /// Accumulates data when native update is not available
-    fn store_data(&mut self, data: &[u8]) -> Result<(), Error> {
-        match &mut self.data {
-            Some(buffer) => {
-                if buffer.len() + data.len() > MAX_BUFFER_LEN {
-                    return Err(Error::new(ErrorKind::BufferSize));
-                }
-                buffer.extend_from_slice(data);
-            }
-            None => {
-                if data.len() > MAX_BUFFER_LEN {
-                    return Err(Error::new(ErrorKind::BufferSize));
-                }
-                self.data = Some(data.to_vec());
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Creates a new message sign context for a specific
-    /// algorithm name, requires a private key and parameters.
-    pub fn message_sign_new(
-        libctx: &OsslContext,
-        alg: SigAlg,
-        key: &mut EvpPkey,
-        params: Option<&OsslParam>,
-    ) -> Result<OsslSignature, Error> {
-        let mut ctx = Self::init(libctx, key)?;
-
-        if sigalg_is_oneshot(alg) {
-            /* Single shot algorithms must always use EVP_PKEY_sign_init */
-            if unsafe { EVP_PKEY_sign_init(ctx.pkey_ctx.as_mut_ptr()) } != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-            if let Some(p) = params {
-                let ret = unsafe {
-                    EVP_PKEY_CTX_set_params(
-                        ctx.pkey_ctx.as_mut_ptr(),
-                        p.as_ptr(),
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-            }
-            return Ok(ctx);
-        }
-
-        if sigalg_uses_legacy_api(alg) {
-            #[cfg(not(feature = "fips"))]
-            {
-                let mut lctx = EvpMdCtx::new()?;
-                let ret = unsafe {
-                    EVP_DigestSignInit_ex(
-                        lctx.as_mut_ptr(),
-                        std::ptr::null_mut(),
-                        sigalg_to_digest_ptr(alg),
-                        libctx.ptr(),
-                        std::ptr::null(),
-                        key.as_mut_ptr(),
-                        match params {
-                            Some(p) => p.as_ptr(),
-                            None => std::ptr::null(),
-                        },
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-                ctx.legacy_ctx = Some(lctx);
-            }
-            #[cfg(feature = "fips")]
-            {
-                let mut lctx = fips::ProviderSignatureCtx::new(alg)?;
-                lctx.digest_sign_init(
-                    sigalg_to_digest_ptr(alg),
-                    key,
-                    match params {
-                        Some(p) => p.as_ptr(),
-                        None => std::ptr::null(),
-                    },
-                )?;
-                ctx.legacy_ctx = Some(lctx);
-            }
-            ctx.supports_updates = match sigalg_supports_updates(alg) {
-                Some(true) => true,
-                Some(false) => false,
-                None => false,
-            };
-            return Ok(ctx);
-        }
-
-        #[cfg(ossl_v350)]
-        {
-            let ret = unsafe {
-                EVP_PKEY_sign_message_init(
-                    ctx.pkey_ctx.as_mut_ptr(),
-                    EvpSignature::new(libctx, sigalg_to_ossl_name(alg))?
-                        .as_mut_ptr(),
-                    match params {
-                        Some(p) => p.as_ptr(),
-                        None => std::ptr::null(),
-                    },
-                )
-            };
-            if ret != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-            ctx.supports_updates = match sigalg_supports_updates(alg) {
-                Some(true) => true,
-                Some(false) => false,
-                None => {
-                    /* OpenSSL 3.5 implements only one shot ML-DSA,
-                     * while later implementations can deal with
-                     * update()/final() operations. Probe here, and
-                     * set up a backup buffer if update()s are not
-                     * supported.
-                     */
-                    let ret = unsafe {
-                        EVP_PKEY_sign_message_update(
-                            ctx.pkey_ctx.as_mut_ptr(),
-                            std::ptr::null(),
-                            0,
-                        )
-                    };
-                    ret == 1
-                }
-            };
-            return Ok(ctx);
-        }
-
-        #[cfg(not(ossl_v350))]
-        Err(Error::new(ErrorKind::WrapperError))
-    }
-
-    /// One shot signature, takes data and a buffer where to store
-    /// the signature. The signature buffer must have enough space
-    /// to receive the signature. On success the siganture length is
-    /// returned.
-    pub fn message_sign(
-        &mut self,
-        data: &[u8],
-        signature: Option<&mut [u8]>,
-    ) -> Result<usize, Error> {
-        let mut siglen = match &signature {
-            Some(sig) => sig.len(),
-            None => 0,
-        };
-        let siglen_ptr: *mut usize = &mut siglen;
-
-        /* check siglen buffer is enough */
-        if let Some(ctx) = &mut self.legacy_ctx {
-            #[cfg(not(feature = "fips"))]
-            {
-                let ret = unsafe {
-                    EVP_DigestSign(
-                        ctx.as_mut_ptr(),
-                        std::ptr::null_mut(),
-                        siglen_ptr,
-                        data.as_ptr(),
-                        data.len(),
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-            }
-            #[cfg(feature = "fips")]
-            {
-                siglen = ctx.digest_sign(None, data)?;
-            }
-        } else {
-            let ret = unsafe {
-                EVP_PKEY_sign(
-                    self.pkey_ctx.as_mut_ptr(),
-                    std::ptr::null_mut(),
-                    siglen_ptr,
-                    data.as_ptr(),
-                    data.len(),
-                )
-            };
-            if ret != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-        }
-
-        if let Some(sig) = signature {
-            if siglen > sig.len() {
-                return Err(Error::new(ErrorKind::BufferSize));
-            }
-
-            if let Some(ctx) = &mut self.legacy_ctx {
-                #[cfg(not(feature = "fips"))]
-                {
-                    let ret = unsafe {
-                        EVP_DigestSign(
-                            ctx.as_mut_ptr(),
-                            sig.as_mut_ptr(),
-                            siglen_ptr,
-                            data.as_ptr(),
-                            data.len(),
-                        )
-                    };
-                    if ret != 1 {
-                        return Err(Error::new(ErrorKind::OsslError));
-                    }
-                }
-                #[cfg(feature = "fips")]
-                {
-                    siglen = ctx.digest_sign(Some(sig), data)?;
-                }
-            } else {
-                let ret = unsafe {
-                    EVP_PKEY_sign(
-                        self.pkey_ctx.as_mut_ptr(),
-                        sig.as_mut_ptr(),
-                        siglen_ptr,
-                        data.as_ptr(),
-                        data.len(),
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-            }
-        }
-
-        Ok(siglen)
-    }
-
-    /// Feeds data to the message sign provider
-    pub fn message_sign_update(&mut self, data: &[u8]) -> Result<(), Error> {
-        if !self.supports_updates {
-            return self.store_data(data);
-        }
-
-        if let Some(ctx) = &mut self.legacy_ctx {
-            #[cfg(not(feature = "fips"))]
-            {
-                let ret = unsafe {
-                    EVP_DigestSignUpdate(
-                        ctx.as_mut_ptr(),
-                        data.as_ptr() as *const c_void,
-                        data.len(),
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-            }
-            #[cfg(feature = "fips")]
-            ctx.digest_sign_update(data)?;
-
-            return Ok(());
-        }
-
-        #[cfg(ossl_v350)]
-        {
-            let ret = unsafe {
-                EVP_PKEY_sign_message_update(
-                    self.pkey_ctx.as_mut_ptr(),
-                    data.as_ptr(),
-                    data.len(),
-                )
-            };
-            if ret != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-
-            return Ok(());
-        }
-
-        #[cfg(not(ossl_v350))]
-        Err(Error::new(ErrorKind::WrapperError))
-    }
-
-    /// Finalizes data and generates signature
-    pub fn message_sign_final(
-        &mut self,
-        signature: &mut [u8],
-    ) -> Result<usize, Error> {
-        if !self.supports_updates {
-            if let Some(buffer) = self.data.take() {
-                return self.message_sign(buffer.as_slice(), Some(signature));
-            }
-
-            return Err(Error::new(ErrorKind::WrapperError));
-        }
-
-        if let Some(ctx) = &mut self.legacy_ctx {
-            #[cfg(not(feature = "fips"))]
-            {
-                let mut siglen = signature.len();
-                let siglen_ptr: *mut usize = &mut siglen;
-
-                let ret = unsafe {
-                    EVP_DigestSignFinal(
-                        ctx.as_mut_ptr(),
-                        signature.as_mut_ptr(),
-                        siglen_ptr,
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-                return Ok(siglen);
-            }
-            #[cfg(feature = "fips")]
-            return ctx.digest_sign_final(signature);
-        }
-
-        #[cfg(ossl_v350)]
-        {
-            let mut siglen = signature.len();
-            let siglen_ptr: *mut usize = &mut siglen;
-
-            let ret = unsafe {
-                EVP_PKEY_sign_message_final(
-                    self.pkey_ctx.as_mut_ptr(),
-                    signature.as_mut_ptr(),
-                    siglen_ptr,
-                )
-            };
-            if ret != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-
-            return Ok(siglen);
-        }
-
-        #[cfg(not(ossl_v350))]
-        Err(Error::new(ErrorKind::WrapperError))
-    }
-
-    /// Creates a new message verify context for a specific
-    /// algorithm name, requires a public key and parameters.
-    pub fn message_verify_new(
-        libctx: &OsslContext,
-        alg: SigAlg,
-        key: &mut EvpPkey,
-        params: Option<&OsslParam>,
-    ) -> Result<OsslSignature, Error> {
-        let mut ctx = Self::init(libctx, key)?;
-
-        if sigalg_is_oneshot(alg) {
-            /* Single shot algorithms must always use EVP_PKEY_verify_init */
-            if unsafe { EVP_PKEY_verify_init(ctx.pkey_ctx.as_mut_ptr()) } != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-            if let Some(p) = params {
-                let ret = unsafe {
-                    EVP_PKEY_CTX_set_params(
-                        ctx.pkey_ctx.as_mut_ptr(),
-                        p.as_ptr(),
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-            }
-            return Ok(ctx);
-        }
-
-        if sigalg_uses_legacy_api(alg) {
-            #[cfg(not(feature = "fips"))]
-            {
-                let mut lctx = EvpMdCtx::new()?;
-                let ret = unsafe {
-                    EVP_DigestVerifyInit_ex(
-                        lctx.as_mut_ptr(),
-                        std::ptr::null_mut(),
-                        sigalg_to_digest_ptr(alg),
-                        libctx.ptr(),
-                        std::ptr::null(),
-                        key.as_mut_ptr(),
-                        match params {
-                            Some(p) => p.as_ptr(),
-                            None => std::ptr::null(),
-                        },
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-                ctx.legacy_ctx = Some(lctx);
-            }
-            #[cfg(feature = "fips")]
-            {
-                let mut lctx = fips::ProviderSignatureCtx::new(alg)?;
-                lctx.digest_verify_init(
-                    sigalg_to_digest_ptr(alg),
-                    key,
-                    match params {
-                        Some(p) => p.as_ptr(),
-                        None => std::ptr::null(),
-                    },
-                )?;
-                ctx.legacy_ctx = Some(lctx);
-            }
-            ctx.supports_updates = match sigalg_supports_updates(alg) {
-                Some(true) => true,
-                Some(false) => false,
-                None => false,
-            };
-            return Ok(ctx);
-        }
-
-        #[cfg(ossl_v350)]
-        {
-            let ret = unsafe {
-                EVP_PKEY_verify_message_init(
-                    ctx.pkey_ctx.as_mut_ptr(),
-                    EvpSignature::new(libctx, sigalg_to_ossl_name(alg))?
-                        .as_mut_ptr(),
-                    match params {
-                        Some(p) => p.as_ptr(),
-                        None => std::ptr::null(),
-                    },
-                )
-            };
-            if ret != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-            ctx.supports_updates = match sigalg_supports_updates(alg) {
-                Some(true) => true,
-                Some(false) => false,
-                None => {
-                    /* OpenSSL 3.5 implements only one shot ML-DSA,
-                     * while later implementations can deal with
-                     * update()/final() operations. Probe here, and
-                     * set up a backup buffer if update()s are not
-                     * supported.
-                     */
-                    let ret = unsafe {
-                        EVP_PKEY_verify_message_update(
-                            ctx.pkey_ctx.as_mut_ptr(),
-                            std::ptr::null(),
-                            0,
-                        )
-                    };
-                    ret == 1
-                }
-            };
-            return Ok(ctx);
-        }
-
-        #[cfg(not(ossl_v350))]
-        Err(Error::new(ErrorKind::WrapperError))
-    }
-
-    /// One shot verification function
-    pub fn message_verify(
-        &mut self,
-        data: &[u8],
-        signature: Option<&[u8]>,
-    ) -> Result<(), Error> {
-        let sig = match &signature {
-            Some(s) => s,
-            None => match &self.signature {
-                Some(v) => v.as_slice(),
-                None => return Err(Error::new(ErrorKind::WrapperError)),
-            },
-        };
-        if let Some(ctx) = &mut self.legacy_ctx {
-            #[cfg(not(feature = "fips"))]
-            {
-                let ret = unsafe {
-                    EVP_DigestVerify(
-                        ctx.as_mut_ptr(),
-                        sig.as_ptr(),
-                        sig.len(),
-                        data.as_ptr(),
-                        data.len(),
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-            }
-            #[cfg(feature = "fips")]
-            {
-                ctx.digest_verify(sig, data)?;
-            }
-        } else {
-            let ret = unsafe {
-                EVP_PKEY_verify(
-                    self.pkey_ctx.as_mut_ptr(),
-                    sig.as_ptr(),
-                    sig.len(),
-                    data.as_ptr(),
-                    data.len(),
-                )
-            };
-            if ret != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-        }
-        Ok(())
-    }
-
-    /// Unsupported in versions of OpenSSL prior to 3.5.0
-    #[cfg(not(ossl_v350))]
-    pub fn set_signature(&mut self, _signature: &[u8]) -> Result<(), Error> {
-        Err(Error::new(ErrorKind::WrapperError))
-    }
-
-    /// Sets the signature for a VerifySignature operation.
-    /// If the OpenSSL backend supports setting it early (via
-    /// `EVP_PKEY_CTX_set_signature`), it does so; otherwise, it stores the
-    /// signature internally for later use.
-    #[cfg(ossl_v350)]
-    pub fn set_signature(&mut self, signature: &[u8]) -> Result<(), Error> {
-        if self.legacy_ctx.is_some() || !self.supports_updates {
-            self.signature = Some(signature.to_vec());
-            return Ok(());
-        }
-
-        let ret = unsafe {
-            EVP_PKEY_CTX_set_signature(
-                self.pkey_ctx.as_mut_ptr(),
-                signature.as_ptr(),
-                signature.len(),
-            )
-        };
-        if ret != 1 {
-            return Err(Error::new(ErrorKind::OsslError));
-        }
-
-        Ok(())
-    }
-
-    /// Feeds data to the message verify provider
-    pub fn message_verify_update(&mut self, data: &[u8]) -> Result<(), Error> {
-        if !self.supports_updates {
-            return self.store_data(data);
-        }
-
-        if let Some(ctx) = &mut self.legacy_ctx {
-            #[cfg(not(feature = "fips"))]
-            {
-                let ret = unsafe {
-                    EVP_DigestVerifyUpdate(
-                        ctx.as_mut_ptr(),
-                        data.as_ptr() as *const c_void,
-                        data.len(),
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                }
-            }
-            #[cfg(feature = "fips")]
-            ctx.digest_sign_update(data)?;
-
-            return Ok(());
-        }
-
-        #[cfg(ossl_v350)]
-        {
-            let ret = unsafe {
-                EVP_PKEY_verify_message_update(
-                    self.pkey_ctx.as_mut_ptr(),
-                    data.as_ptr(),
-                    data.len(),
-                )
-            };
-            if ret != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-
-            return Ok(());
-        }
-
-        #[cfg(not(ossl_v350))]
-        Err(Error::new(ErrorKind::WrapperError))
-    }
-
-    /// Finalizes data and generates signature
-    pub fn message_verify_final(
-        &mut self,
-        signature: Option<&[u8]>,
-    ) -> Result<(), Error> {
-        if !self.supports_updates {
-            if let Some(buffer) = self.data.take() {
-                return self.message_verify(buffer.as_slice(), signature);
-            }
-
-            return Err(Error::new(ErrorKind::WrapperError));
-        }
-
-        if let Some(ctx) = &mut self.legacy_ctx {
-            let sig = match &signature {
-                Some(s) => s,
-                None => match &self.signature {
-                    Some(v) => v.as_slice(),
-                    None => return Err(Error::new(ErrorKind::NullPtr)),
-                },
-            };
-            #[cfg(not(feature = "fips"))]
-            {
-                let ret = unsafe {
-                    EVP_DigestVerifyFinal(
-                        ctx.as_mut_ptr(),
-                        sig.as_ptr(),
-                        sig.len(),
-                    )
-                };
-                if ret != 1 {
-                    return Err(Error::new(ErrorKind::OsslError));
-                } else {
-                    return Ok(());
-                }
-            }
-            #[cfg(feature = "fips")]
-            return ctx.digest_verify_final(sig);
-        }
-
-        #[cfg(ossl_v350)]
-        {
-            match signature {
-                Some(sig) => {
-                    self.set_signature(sig)?;
-                }
-                None => (),
-            }
-
-            let ret = unsafe {
-                EVP_PKEY_verify_message_final(self.pkey_ctx.as_mut_ptr())
-            };
-            if ret != 1 {
-                return Err(Error::new(ErrorKind::OsslError));
-            }
-
-            return Ok(());
-        }
-
-        #[cfg(not(ossl_v350))]
-        Err(Error::new(ErrorKind::WrapperError))
-    }
-}
